refactor: extract shared node execution interfaces to nodes/types.ts
- Create nodes/types.ts with NodeExecutionContext and NodeExecutionResult interfaces
- Remove duplicated interfaces from HttpNode.service.ts and EmailNode.service.ts
- Update imports in HttpNode.test.ts and EmailNode.test.ts to use shared types
- Clean up nodes/index.ts to re-export shared types properly
- Fix TypeScript export issues with 'export type' for isolated modules
- All tests passing (26/26 HttpNode, 15/15 EmailNode, 7/7 node-definitions)
- Build successful with no TypeScript errors

This eliminates code duplication and provides a single source of truth for node execution interfaces.
